[Portal] Give CIMD and DCR their own pages under Applications
One tab holding both mechanisms was 7 cards and ~20 controls, with the
token-lifetime block appearing twice, because CIMD and DCR are
configured independently and their settings are structurally identical
-- so the page read as duplicated. Each is now a nav page of its own
under Applications (configuration/apps/metadata-documents and
.../self-registration), and the tab keeps only what the two share.

Each page owns its own useAppConfigForm covering just its own config
section. That is the substantive change, not the layout: until now a
save from any tab committed the whole Applications form, so an
uncommitted edit to one mechanism rode along with a save of the other.
ApplicationsConfigurationScreen's FormState is back to { clients }.

The tab keeps the client counts, the allowed-API-resources card, and an
accent-tinted card -- the Danger zone card's geometry in the accent
scale -- naming each mechanism, its state, and a way to its page. Its
body text is deliberately neutral: --accent-11 is the colour Link
renders in, so tinted prose there read as a link.

The client list is reachable from the counts card with ?source
preselecting its filter, and its route is renamed from apps/dcr to
apps/dynamic-clients, which is what it has listed since CIMD rows
joined it.

Trusted domains follows the Endpoint Direct Access radio-group idiom
with the domain list nested under the restricted option, so the list
belongs to the choice structurally rather than by proximity.

ref DEV-3838
