+++ portal/src/components/dynamic-clients/ClientCountCard.module.css
@@ -1,0 +1,17 @@
+.stats {
+  @apply flex flex-row flex-wrap gap-x-10 gap-y-4;
+}
+
+.stat {
+  @apply flex flex-col gap-1;
+  min-width: 4rem;
+}
+
+.statLabel,
+.statValue {
+  margin: 0;
+}
+
+.statValue {
+  color: var(--gray-12);
+}
