+++ portal/src/components/dynamic-clients/DynamicClientsTab.module.css
@@ -2,24 +2,82 @@
   @apply flex flex-col gap-4;
 }
 
-.rootWithSaveBar {
-  padding-bottom: 5.5rem;
+/* The Danger zone card's geometry (UserDetailsAccountStatus.module.css),
+   tinted with the accent scale instead of the red one: same 200px title
+   column, same 24px padding and 12px radius, same divider-separated rows
+   with the action held to the right until the tablet breakpoint. */
+.mechanismCard {
+  display: grid;
+  grid-template-columns: 200px minmax(0, 1fr);
+  gap: 48px;
+  align-items: start;
+  padding: 24px;
+  border: 1px solid var(--accent-7);
+  border-radius: 12px;
+  background-color: var(--accent-a2);
+
+  @apply tablet:grid-cols-1 tablet:gap-4;
 }
 
-.stats {
-  @apply flex flex-row flex-wrap gap-x-10 gap-y-4;
-}
-
-.stat {
-  @apply flex flex-col gap-1;
-  min-width: 4rem;
-}
-
-.statLabel,
-.statValue {
+.mechanismCardTitle {
   margin: 0;
+  color: var(--accent-12);
 }
 
-.statValue {
-  color: var(--gray-12);
+.mechanismCardContent {
+  min-width: 0;
+}
+
+.mechanismCardContent > :first-child {
+  padding-top: 0;
+}
+
+.mechanismCardContent > :last-child {
+  padding-bottom: 0;
+  border-bottom: 0;
+}
+
+.mechanismRow {
+  grid-template:
+    "label  action"
+    "body   action"
+    "status action"
+    / minmax(0, 1fr) max-content;
+  @apply grid;
+  @apply gap-x-6 gap-y-1;
+  @apply py-6;
+  border-bottom: 1px solid var(--gray-5);
+
+  /* Below tablet the button no longer fits beside the text without crushing
+     it, so everything stacks -- as the Danger zone card does. */
+  @apply tablet:flex tablet:flex-col tablet:gap-2;
+}
+
+.mechanismRowLabel {
+  grid-area: label;
+  margin: 0;
+  line-height: 20px;
+  font-weight: 600;
+  color: var(--accent-12);
+}
+
+.mechanismRowBody {
+  grid-area: body;
+  margin: 0;
+  line-height: 20px;
+  max-width: 500px;
+  /* Neutral, not accent: --accent-11 is the colour Link renders in, so tinted
+     body text here read as a link. The card's tint and border carry the
+     grouping; the prose does not need to. */
+  color: var(--gray-11);
+}
+
+.mechanismRowStatus {
+  grid-area: status;
+  @apply flex flex-row items-center pt-1;
+}
+
+.mechanismRowAction {
+  grid-area: action;
+  @apply self-center;
 }
