--- portal/src/components/dynamic-clients/MechanismSectionHeader.module.css
@@ -1,27 +1,0 @@
-.header {
-  @apply flex flex-row items-start justify-between gap-6;
-  @apply pt-2;
-}
-
-.headingGroup {
-  @apply flex flex-col gap-1 min-w-0;
-}
-
-.title,
-.description {
-  margin: 0;
-}
-
-.title {
-  color: var(--gray-12);
-}
-
-.toggle {
-  @apply flex-none pt-1;
-}
-
-@media (max-width: 40rem) {
-  .header {
-    @apply flex-col gap-3;
-  }
-}
