+++ portal/src/locale-data/en.json
@@ -153,6 +153,12 @@
   "ScreenNav.external-oauth": "Social / Enterprise Login",
   "ScreenNav.login-methods": "Login Methods",
   "ScreenNav.client-applications": "Applications",
+  "ScreenNav.cimd": "CIMD",
+  "ScreenNav.dcr": "DCR",
+  "MetadataDocumentsScreen.title": "Client ID Metadata Document (CIMD) Registration",
+  "MetadataDocumentsScreen.enable.title": "Client ID metadata documents",
+  "SelfRegistrationScreen.title": "Dynamic Client Registration (DCR)",
+  "SelfRegistrationScreen.enable.title": "Dynamic client registration",
   "ScreenNav.localization": "Email/SMS Templates",
   "ScreenNav.branding": "Branding",
   "ScreenNav.design": "Design",
@@ -803,14 +809,15 @@
   "PasswordSettings.resetPasswordWithPhone.options.sms": "SMS",
   "PasswordSettings.resetPasswordWithPhone.options.whatsapp": "WhatsApp",
   "PasswordSettings.resetPasswordWithPhone.options.whatsappOrSMS": "WhatsApp or SMS",
-  "ApplicationsConfigurationScreen.title": "Applications",
+  "ApplicationsConfigurationScreen.title": "All applications",
   "ApplicationsConfigurationScreen.description": "Configure applications to use this project for authentication.",
   "ApplicationsConfigurationScreen.tab.applications": "Applications",
   "ApplicationsConfigurationScreen.tab.dynamic-clients": "Dynamic clients",
   "DynamicClientsTab.enable.toggle.label": "Enable dynamic client registration",
   "DynamicClientsTab.enable.description": "Clients such as <mcpLink>MCP clients</mcpLink> register themselves at an endpoint (RFC 7591). <dcrLink>Learn more.</dcrLink>",
-  "DynamicClientsTab.section.registration.title": "Dynamic client registration (DCR)",
-  "DynamicClientsTab.clients.title": "Clients",
+  "DynamicClientsTab.clients.title": "Dynamic Clients",
+  "DynamicClientsTab.clients.source.cimd": "Via CIMD",
+  "DynamicClientsTab.clients.source.dcr": "Via DCR",
   "DynamicClientsTab.clients.view-all": "View all",
   "DynamicClientListScreen.title": "Dynamic clients",
   "DynamicClientsTab.registration-endpoint.label": "Registration endpoint",
@@ -3060,14 +3067,11 @@
   "HookConfigurationScreen.non-blocking-handlers.empty": "No non-blocking hooks yet.",
   "HookConfigurationScreen.non-blocking-handler.name.label": "Name",
   "HookConfigurationScreen.non-blocking-handler.name.placeholder": "e.g. Send welcome email",
-  "CIMDSection.title": "Client ID metadata documents (CIMD)",
   "CIMDSection.enable.description": "Apps identify themselves with a metadata document they host, fetched on first sign-in. <mcpLink>MCP clients</mcpLink> prefer it to registration.",
   "CIMDSection.enable.toggle.label": "Enable client ID metadata documents",
   "CIMDSection.trusted-domains.title": "Trusted domains",
   "CIMDSection.trusted-domains.any.title": "Any domain",
-  "CIMDSection.trusted-domains.any.subtitle": "Any app with a valid document can sign in.",
-  "CIMDSection.trusted-domains.list.title": "Only these domains",
-  "CIMDSection.trusted-domains.list.subtitle": "Only apps whose document is on a listed domain.",
+  "CIMDSection.trusted-domains.list.title": "Only specific domains",
   "CIMDSection.trusted-domains.label": "Domains",
   "CIMDSection.trusted-domains.wildcard-hint": "An exact hostname, or a wildcard such as *.example.com.",
   "CIMDSection.trusted-domains.placeholder": "mcp.example.com",
@@ -3095,5 +3099,7 @@
   "DynamicClientsTab.clients.count.unknown": "—",
   "DynamicClientsTab.registration-endpoint.description": "Clients POST their metadata here to register (RFC 7591).",
   "DynamicClientsTab.client-config.title": "Client configuration",
-  "DynamicClientsTab.client-config.description": "Token settings for every DCR-registered client."
+  "DynamicClientsTab.client-config.description": "Token settings for every DCR-registered client.",
+  "DynamicClientsTab.configure": "Configure",
+  "DynamicClientsTab.mechanisms.title": "Onboarding mechanisms"
 }
